move, rename, and enhance HOC by adding venueEntityLoaded flag
